Refactor route components
